tmp/challenge/medium_honey_i_shrunk_the_pets/requirements.txt
Pillow==9.0.1
flask
gunicorn
markdown
